requirements.txt
protobuf==5.29.1
grpcio==1.68.1
grpcio-tools==1.68.1
networkx==3.0
